Applied Extract class refactoring and created new class SqlCastManager
